document how to install "rc.local" to bypass broken AX210 firmware
In e4bd4c9 we document how to manually remove the broken
"iwlwifi-ty-a0-gf-a0.pnvm.xz" firmware from the Fedora template.  Here
we script *bypassing* this firmware when sys-net boots, pending either
(a) Salting this workaround in a future securedrop-workstation release
or (b) an upstream fix to this incompatibility with the AX210 firmware.