Add project status page and create overview section
This PR slightly reorganizes the content, pulling out
general purpose information into a new Overview section
at the top of the table of contents.

Additionally, and most substantively, this adds a new
project status page, which highlights the current
status of the project, a brief roadmap and timeline,
and information about circumstances where SecureDrop
Workstation is likely to provide the best experience.