Fixes #212 by breaking up the install guide into a new section.
This commit is a slight reorganization of the Admin Guide into
two sections, similar to how SecureDrop (server) docs are done.
With these changes, there is now one section dedicated to the
installation, and one section to general reference for Admins.

Along with the changes, the individual pages, screenshots, and
images are reorganized, and internal references have been updated
where necessary.