Overhaul the hardware guide
This commit overhauls the hardware guide by making a few key
changes:

1. It removes our specific laptop recommendations, instead
referring people to the Qubes hardware compatibility list.
2. It clarifies a few specific requirements for hardware, namely
32GB of RAM (removing the confusing language about production use),
as well as a USB 3.0 flash drive with minimum 8GB of capacity.
3. It provides a generalized BIOS update section, which points
folks to fwupd on supported systems, and to their manufacturers
instructions for instances where fwupd is not supported.